+++ app/models/product_option.rb
@@ -1,0 +1,7 @@
+class ProductOption < ApplicationRecord
+  belongs_to :product, optional: true
+  belongs_to :product_variant, optional: true
+  belongs_to :product_option, optional: true
+
+  nilify_blanks only: [:option]
+end