+++ app/presenters/bookmark_presenter.rb
@@ -10,4 +10,8 @@ class BookmarkPresenter < ItemPresenter
   def delete_confirm_msg
     I18n.t('remove_bookmark.confirm', name: display_name)
   end
+
+  def product_option
+    nil
+  end
 end