+++ db/migrate/20240819195526_create_product_options.rb
@@ -1,0 +1,11 @@
+class CreateProductOptions < ActiveRecord::Migration[7.1]
+  def change
+    create_table :product_options do |t|
+      t.string :option
+      t.bigint :product_id
+      t.bigint :product_variant_id
+
+      t.timestamps
+    end
+  end
+end