+++ db/migrate/20240819205747_add_index_to_product_options.rb
@@ -1,0 +1,6 @@
+class AddIndexToProductOptions < ActiveRecord::Migration[7.1]
+  def change
+    add_index :product_options, [:option, :product_id], unique: true
+    add_index :product_options, [:option, :product_variant_id], unique: true
+  end
+end