+++ db/migrate/20240820155006_add_not_null_to_product_options.rb
@@ -1,0 +1,5 @@
+class AddNotNullToProductOptions < ActiveRecord::Migration[7.1]
+  def change
+    change_column :product_options, :option, :string, null: false
+  end
+end