+++ db/migrate/20240820194334_add_product_option_to_possession.rb
@@ -1,0 +1,5 @@
+class AddProductOptionToPossession < ActiveRecord::Migration[7.1]
+  def change
+    add_column :possessions, :product_option_id, :bigint
+  end
+end